Daejun

+++ backend/src/main/java/teamproject/backend/foodCategory/FoodCategoryServiceImpl.java
@@ -2,8 +2,11 @@ package teamproject.backend.foodCategory;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import teamproject.backend.domain.FoodCategory;
 import teamproject.backend.foodCategory.dto.FoodCategoryResponse;
+import teamproject.backend.response.BaseException;
+import teamproject.backend.response.BaseExceptionStatus;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -32,17 +35,29 @@ public class FoodCategoryServiceImpl implements FoodCategoryService {
     @Override
     public Long save(String foodCategoryName) {
         //중복 검사
+        FoodCategory foodCategory = foodCategoryRepository.findByCategoryName(foodCategoryName);
+        if(foodCategory != null) throw new BaseException(BaseExceptionStatus.EXIST_CATEGORY);
 
-        FoodCategory foodCategory = new FoodCategory(foodCategoryName);
-        foodCategoryRepository.save(foodCategory);
-        return foodCategory.getCategory_id();
+        FoodCategory newFoodCategory = new FoodCategory(foodCategoryName);
+        foodCategoryRepository.save(newFoodCategory);
+        return newFoodCategory.getCategory_id();
     }
 
     @Override
+    @Transactional
     public void delete(String foodCategoryName) {
         //존재 여부 찾기
-
         FoodCategory foodCategory = foodCategoryRepository.findByCategoryName(foodCategoryName);
+        if(foodCategory == null) throw new BaseException(BaseExceptionStatus.NOT_EXIST_CATEGORY);
+
         foodCategoryRepository.delete(foodCategory);
+    }
+
+    @Override
+    @Transactional
+    public void change(String before, String after) {
+        FoodCategory foodCategory = foodCategoryRepository.findByCategoryName(before);
+
+        foodCategory.setCategoryName(after);
     }
 }
